Issue #19: Add bundle name to base driver
Currently hyperv and libvirt have this as a driver specific field, but
this was forgotten in hyperkit. This will be needed to be able to know
which version of CRC is running from a ~/.crc/machines/crc directory.

We limit ourselves to a "BundleName" rather than a "BundePath" as we
cannot rely on files outside of ~/.crc to be untouched by the user (ie
not moved/removed). Since bundles are extracted to
~/.crc/cache/$BundleName, this the name is going to be enough to find
back the bundle content.

https://github.com/code-ready/machine/issues/19